Separating code into multiple files to avoid lines over 400
